# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RightTemp OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


export const LeadStatus = {
  new: 'new',
  contacted: 'contacted',
  qualified: 'qualified',
  proposal: 'proposal',
  won: 'won',
  lost: 'lost',
} as const;

export type LeadSource = typeof LeadSource[keyof typeof LeadSource];


export const LeadSource = {
  website: 'website',
  referral: 'referral',
  phone: 'phone',
  walk_in: 'walk_in',
  social_media: 'social_media',
  other: 'other',
} as const;

export interface Lead {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  status: LeadStatus;
  source: LeadSource;
  /** @nullable */
  serviceType?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type LeadInputStatus = typeof LeadInputStatus[keyof typeof LeadInputStatus];


export const LeadInputStatus = {
  new: 'new',
  contacted: 'contacted',
  qualified: 'qualified',
  proposal: 'proposal',
  won: 'won',
  lost: 'lost',
} as const;

export type LeadInputSource = typeof LeadInputSource[keyof typeof LeadInputSource];


export const LeadInputSource = {
  website: 'website',
  referral: 'referral',
  phone: 'phone',
  walk_in: 'walk_in',
  social_media: 'social_media',
  other: 'other',
} as const;

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone: string;
  address?: string;
  status: LeadInputStatus;
  source: LeadInputSource;
  serviceType?: string;
  notes?: string;
}

export type LeadUpdateStatus = typeof LeadUpdateStatus[keyof typeof LeadUpdateStatus];


export const LeadUpdateStatus = {
  new: 'new',
  contacted: 'contacted',
  qualified: 'qualified',
  proposal: 'proposal',
  won: 'won',
  lost: 'lost',
} as const;

export type LeadUpdateSource = typeof LeadUpdateSource[keyof typeof LeadUpdateSource];


export const LeadUpdateSource = {
  website: 'website',
  referral: 'referral',
  phone: 'phone',
  walk_in: 'walk_in',
  social_media: 'social_media',
  other: 'other',
} as const;

export interface LeadUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  status?: LeadUpdateStatus;
  source?: LeadUpdateSource;
  serviceType?: string;
  notes?: string;
}

export type CustomerStatus = typeof CustomerStatus[keyof typeof CustomerStatus];


export const CustomerStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type CustomerServiceType = typeof CustomerServiceType[keyof typeof CustomerServiceType];


export const CustomerServiceType = {
  hvac_install: 'hvac_install',
  hvac_repair: 'hvac_repair',
  maintenance: 'maintenance',
  inspection: 'inspection',
  emergency: 'emergency',
  other: 'other',
} as const;

export interface Customer {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  status: CustomerStatus;
  serviceType: CustomerServiceType;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CustomerInputStatus = typeof CustomerInputStatus[keyof typeof CustomerInputStatus];


export const CustomerInputStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type CustomerInputServiceType = typeof CustomerInputServiceType[keyof typeof CustomerInputServiceType];


export const CustomerInputServiceType = {
  hvac_install: 'hvac_install',
  hvac_repair: 'hvac_repair',
  maintenance: 'maintenance',
  inspection: 'inspection',
  emergency: 'emergency',
  other: 'other',
} as const;

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone: string;
  address?: string;
  status: CustomerInputStatus;
  serviceType: CustomerInputServiceType;
  notes?: string;
}

export type CustomerUpdateStatus = typeof CustomerUpdateStatus[keyof typeof CustomerUpdateStatus];


export const CustomerUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type CustomerUpdateServiceType = typeof CustomerUpdateServiceType[keyof typeof CustomerUpdateServiceType];


export const CustomerUpdateServiceType = {
  hvac_install: 'hvac_install',
  hvac_repair: 'hvac_repair',
  maintenance: 'maintenance',
  inspection: 'inspection',
  emergency: 'emergency',
  other: 'other',
} as const;

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  status?: CustomerUpdateStatus;
  serviceType?: CustomerUpdateServiceType;
  notes?: string;
}

export type DashboardStatsLeadsByStatus = {
  new: number;
  contacted: number;
  qualified: number;
  proposal: number;
  won: number;
  lost: number;
};

export interface DashboardStats {
  totalLeads: number;
  totalCustomers: number;
  activeCustomers: number;
  leadsByStatus: DashboardStatsLeadsByStatus;
  recentLeads: Lead[];
  recentCustomers: Customer[];
}

export type GetLeadsParams = {
status?: GetLeadsStatus;
search?: string;
};

export type GetLeadsStatus = typeof GetLeadsStatus[keyof typeof GetLeadsStatus];


export const GetLeadsStatus = {
  new: 'new',
  contacted: 'contacted',
  qualified: 'qualified',
  proposal: 'proposal',
  won: 'won',
  lost: 'lost',
} as const;

export type GetCustomersParams = {
status?: GetCustomersStatus;
search?: string;
};

export type GetCustomersStatus = typeof GetCustomersStatus[keyof typeof GetCustomersStatus];


export const GetCustomersStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;