# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RightTemp OS API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Lead management
  - name: customers
    description: Customer management
  - name: dashboard
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [new, contacted, qualified, proposal, won, lost]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Bad request

  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Bad request

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Lead:
      type: object
      required: [id, name, email, phone, status, source, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, qualified, proposal, won, lost]
        source:
          type: string
          enum: [website, referral, phone, walk_in, social_media, other]
        serviceType:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    LeadInput:
      type: object
      required: [name, email, phone, status, source]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, proposal, won, lost]
        source:
          type: string
          enum: [website, referral, phone, walk_in, social_media, other]
        serviceType:
          type: string
        notes:
          type: string

    LeadUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, proposal, won, lost]
        source:
          type: string
          enum: [website, referral, phone, walk_in, social_media, other]
        serviceType:
          type: string
        notes:
          type: string

    Customer:
      type: object
      required: [id, name, email, phone, status, serviceType, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        serviceType:
          type: string
          enum: [hvac_install, hvac_repair, maintenance, inspection, emergency, other]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    CustomerInput:
      type: object
      required: [name, email, phone, status, serviceType]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive]
        serviceType:
          type: string
          enum: [hvac_install, hvac_repair, maintenance, inspection, emergency, other]
        notes:
          type: string

    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive]
        serviceType:
          type: string
          enum: [hvac_install, hvac_repair, maintenance, inspection, emergency, other]
        notes:
          type: string

    DashboardStats:
      type: object
      required: [totalLeads, totalCustomers, leadsByStatus, activeCustomers, recentLeads, recentCustomers]
      properties:
        totalLeads:
          type: integer
        totalCustomers:
          type: integer
        activeCustomers:
          type: integer
        leadsByStatus:
          type: object
          properties:
            new:
              type: integer
            contacted:
              type: integer
            qualified:
              type: integer
            proposal:
              type: integer
            won:
              type: integer
            lost:
              type: integer
          required: [new, contacted, qualified, proposal, won, lost]
        recentLeads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        recentCustomers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"