# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RightTemp OS API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerServiceType } from './customerServiceType';
import type { CustomerStatus } from './customerStatus';

export interface Customer {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  status: CustomerStatus;
  serviceType: CustomerServiceType;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}