# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RightTemp OS API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerUpdateServiceType } from './customerUpdateServiceType';
import type { CustomerUpdateStatus } from './customerUpdateStatus';

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  status?: CustomerUpdateStatus;
  serviceType?: CustomerUpdateServiceType;
  notes?: string;
}