# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RightTemp OS API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadInputSource } from './leadInputSource';
import type { LeadInputStatus } from './leadInputStatus';

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone: string;
  address?: string;
  status: LeadInputStatus;
  source: LeadInputSource;
  serviceType?: string;
  notes?: string;
}