# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RightTemp OS API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadUpdateSource } from './leadUpdateSource';
import type { LeadUpdateStatus } from './leadUpdateStatus';

export interface LeadUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  status?: LeadUpdateStatus;
  source?: LeadUpdateSource;
  serviceType?: string;
  notes?: string;
}